requirements.txt
dash-cytoscape==0.1.1
dash-daq==0.1.4
dash==1.0.2
dash_bootstrap_components==0.7.0
dash_callback_chain==0.0.2
dash_core_components==1.0.0
dash_html_components==1.0.0
dash_table==4.0.2
dill==0.3.0
feather-format==0.4.0
flask==1.0.3
flask_login==0.4.1
flask_sqlalchemy==2.4.0
flask-wtf==0.14.2
flask_caching==1.7.2
Flask-Mail==0.9.1
fuzzywuzzy==0.17.0
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-api-python-client==1.7.11
gunicorn==19.9.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.0
oauth2client==4.1.3
pandas==0.25.0
peakutils==1.3.2
plotly==4.0.0
praw==6.3.1
pycountry==19.7.15
pygraphviz==1.5
pytest-cov==2.7.1
python-dotenv==0.10.3
python-Levenshtein==0.12.0
python-twitter==3.5
quandl==3.4.8
redis==3.3.4
requests==2.22.0
scikit_learn==0.21.3
sd_material_ui==3.1.2
selenium==3.141.0
spotipy==2.4.4
sqlalchemy==1.3.5
sympy==1.4
textblob==0.15.3
visdcc==0.0.40
werkzeug==0.15.4
wordcloud==1.5.0
xgboost==0.90
